Move FunctionBrowser to separate JSX file
